pyproject.toml
[tool.poetry]
name = "micedice"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.0"
PyYAML = "^5.3.1"
pygsheets = {git = "git@github.com:nithinmurali/pygsheets.git"}
aiosqlite = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
